+++ src/types/ToArrayStringArgs.ts
@@ -22,42 +22,13 @@
  * SOFTWARE.
  */
 
-import 'reflect-metadata';
-import { mergeSchemas } from '@graphql-tools/schema';
-import { buildSchemaSync, registerEnumType } from 'type-graphql';
-import { buildAuthDirective } from '../../src';
-import {
-  authFn,
-  Context,
-  UserRoles,
-  UserPermissions,
-  main
-} from '../__commons';
-import { UserResolver } from './UserResolver';
-import { PostResolver } from './PostResolver';
-
-// Build auth directive
-const authDirective = buildAuthDirective<Context, UserRoles, UserPermissions>({
-  auth: authFn,
-  roles: { enumName: 'UserRoles' },
-  permissions: { enumName: 'UserPermissions' }
-});
-// const authDirective = buildAuthDirective({ auth: authFnClass, ... });
-
-// Register UserRoles enum
-registerEnumType(UserRoles, { name: 'UserRoles' });
-// Register UserPermissions enum
-registerEnumType(UserPermissions, { name: 'UserPermission' });
-
-// Build schema
-let schema = buildSchemaSync({
-  resolvers: [UserResolver, PostResolver]
-});
-schema = mergeSchemas({
-  schemas: [schema],
-  typeDefs: [authDirective.typeDefs]
-});
-schema = authDirective.transformer(schema);
-
-// Main
-main(schema);
+export type ToArrayStringArgs<T> = {
+  /**
+   * Value to flatten.
+   */
+  value: T | T[];
+  /**
+   * Flag if value is enum.
+   */
+  isEnum?: boolean;
+};