+++ src/utils/index.ts
@@ -23,3 +23,4 @@
  */
 
 export * from './container';
+export * from './toArrayString';