+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog and release notes
 
+## v0.3.0
+
+- **Breaking**: Renamed `typeName` to `enumName` in `roles` and `permissions` configurations
+
+- **Breaking**: Rename `defaultValue` to `default` in `roles` and `permissions` configurations
+
+- `roles` and `permissions` configuration `default` value `TRole` value or `TRole` array
+
 ## v0.2.0
 
 ### Features